// src/application/dtos/item/create-item-dtos/create-item-output-dto.ts
import { Either } from "../../../../common/helpers/Either";
import DefaultApplicationError from "../../../errors/application/default-application-error";

type ItemData = {
	id: string;
	userId: string;
	name: string;
	description: string;
	price: number;
	category: string;
	status: string;
};

export type CreateItemOutputDto = Either<DefaultApplicationError, ItemData>;